refactored timer, add timer context
