+++ raynian-fe/src/App.js
@@ -3,6 +3,7 @@ import { FaCog } from "react-icons/fa";
 import Profile from "./components/Profile";
 import { ModalProvider } from "./context/ModalContext";
 import { DarkLightProvider } from "./context/DarkLightContext";
+import { TimerProvider } from "./context/TimerContext";
 import sushi from "./assets/temp_pfp/sushi.jpg";
 import boba from "./assets/temp_pfp/boba.jpg";
 import avocado from "./assets/temp_pfp/avocado.jpg";
@@ -23,7 +24,9 @@ function App() {
           <div className="h-[60%] flex items-center justify-center flex-col">
             <div className="text-center ">
               <p className="text-3xl">You're doing great!</p>
-              <Timer />
+              <TimerProvider>
+                <Timer />
+              </TimerProvider>
               <p className="text-3xl">Sessions completed today:</p>
               <span className="text-3xl">2</span>
             </div>
